display minimal graph with quality property nodes
Removed the previously existing page with an initial view in the
homepage. The idea is that graph interaction is already possible in the
homepage with a limited view. Later, the user can open the graph on a
dedicated page to explore it more. Additionally, esbuild was added as a
bundler for better DX regarding graph implementation.