(wip) introduce some structure and try to render fullscreen graph
When the user clicks on the graph canvas, it should be navigated to
another page where we display the full graph with all quality nodes. On
this page the user can fully interact with the graph. This is where we
also want to add another component to tweak the graph (e.g. force, size,
etc.)